requirements.txt
# AI运输报价系统依赖包列表

# Web框架
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
email-validator==2.1.0.post1

# 数据库
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL驱动
asyncpg==0.29.0  # 异步PostgreSQL驱动
pymongo==4.6.1  # MongoDB驱动

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# 缓存
redis==5.0.1
cachetools==5.3.2

# 数据处理
pandas==2.2.0
numpy==1.26.3
pyarrow==14.0.2

# 机器学习
scikit-learn==1.4.0
joblib==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# LLM集成
openai==1.12.0
langchain==0.1.4
langchain-openai==0.0.5
tiktoken==0.5.2

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.2

# HTTP客户端
httpx==0.26.0
requests==2.31.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
faker==22.5.0

# 文档
pyyaml==6.0.1
jinja2==3.1.3

# 工具
python-slugify==8.0.1
tenacity==8.2.3  # 重试机制
pycountry==23.12.11  # 国家和货币数据
python-dateutil==2.8.2
pytz==2023.3.post1

# 异步支持
aiofiles==23.2.1
aiocache==0.12.2

# 邮件
aiosmtplib==3.0.1

# 开发工具
black==24.1.1  # 代码格式化
isort==5.13.2  # 导入排序
flake8==7.0.0  # 代码检查
mypy==1.8.0  # 类型检查
pre-commit==3.6.0  # Git钩子

# 可选依赖
# dash==2.14.2  # 数据可视化Web应用
# plotly==5.18.0  # 交互式图表
# gunicorn==21.2.0  # 生产环境WSGI服务器 